+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `ui/Scroller` to calculate correctly targetX and targetY when scrolling to a position
+
 ## [4.9.8] - 2025-04-24
 
 No significant changes.